# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NusaAlert API",
        "description": "API dokumentasi untuk NusaAlert.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Daftar alert user",
                "operationId": "3937475256ed06091ff188b4776dace9",
                "parameters": [
                    {
                        "name": "jenis",
                        "in": "query",
                        "description": "Filter berdasarkan jenis bencana",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tanggal",
                        "in": "query",
                        "description": "Filter berdasarkan tanggal",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sent",
                                "read",
                                "dismissed"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar alert berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/alerts/unread": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Daftar alert yang belum dibaca",
                "operationId": "c4d292c8159d80874d291b6d3f62572f",
                "responses": {
                    "200": {
                        "description": "Daftar alert unread"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/alerts/{id}/read": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Tandai alert sebagai sudah dibaca",
                "operationId": "8fbfb54a55293bf48e21b4c78c816296",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert ditandai dibaca"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/alerts/mark-all-read": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Tandai semua alert sebagai sudah dibaca",
                "operationId": "079b4d4886428e09d1dd879efa2ae0b5",
                "responses": {
                    "200": {
                        "description": "Semua alert ditandai dibaca"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
...
}